+++ lib/slack_invite_automation.rb
@@ -1,0 +1,5 @@
+$LOAD_PATH.unshift File.dirname(__FILE__)
+
+module SlackInviteAutomation
+  autoload :Server, 'slack_invite_automation/server'
+end
