+++ lib/slack_invite_automation/views/index.erb
