+++ lib/slack_invite_automation/views/invite.erb
