+++ lib/slack_invite_automation/views/layout.erb
